Move sync client/token handling to separate hook
